Moved HoloViews into own module and added Links support
